requirements.txt
torch
torchvision
transformers
gradio
opencv-python
Pillow
numpy